refactor: Move Loading, Tagline, and EventBus to aux-components
