+++ src/aux-components/components/Loading/Loading.css
