+++ src/aux-components/components/Loading/Loading.ts
@@ -1,7 +1,5 @@
 import Component from 'vue-class-component';
 import Vue from 'vue';
-import { LoadingProgress } from '@casual-simulation/aux-common/LoadingProgress';
-import { appManager, AppType } from '../../../shared/AppManager';
 import { hasValue } from '@casual-simulation/aux-common';
 import { Prop } from 'vue-property-decorator';
 import Tagline from '../Tagline/Tagline';
@@ -16,19 +14,12 @@ export default class Loading extends Vue {
     @Prop({ default: 0 }) progress: number;
     @Prop({ default: null }) error: string;
     @Prop({ default: false }) show: boolean;
+    @Prop({}) version: string;
 
     showSpinner: boolean;
 
     get hasError(): boolean {
         return hasValue(this.error);
-    }
-
-    get isPlayer(): boolean {
-        return appManager.appType === AppType.Player;
-    }
-
-    get version() {
-        return appManager.version.latestTaggedVersion;
     }
 
     onErrorDismiss() {
