+++ src/aux-components/components/Loading/Loading.vue
@@ -31,6 +31,6 @@
         </md-dialog>
     </div>
 </template>
-<script src="./Loading.ts"></script>
+<script src="./Loading.js"></script>
 <style src="./Loading.css" scoped></style>
 <style src="./Loading.global.css"></style>
