+++ src/aux-components/components/Tagline/Tagline.vue
@@ -1,5 +1,5 @@
 <template>
     <div class="tagline">☑️ CasualOS</div>
 </template>
-<script src="./Tagline.ts"></script>
+<script src="./Tagline.js"></script>
 <style src="./Tagline.css" scoped></style>
