+++ src/aux-components/index.ts
@@ -1,1 +1,6 @@
 export * from './SVGPolyfill';
+export * from './EventBus';
+import Loading from './components/Loading/Loading';
+import Tagline from './components/Tagline/Tagline';
+
+export { Loading, Tagline };
