+++ src/aux-server/aux-web/shared/scene/Game.ts
@@ -41,7 +41,7 @@ import { TweenCameraToOperation } from '../interaction/TweenCameraToOperation';
 import { baseAuxAmbientLight, baseAuxDirectionalLight } from './SceneUtils';
 import { createHtmlMixerContext, disposeHtmlMixerContext } from './HtmlUtils';
 import { flatMap, merge, union } from 'lodash';
-import { EventBus } from '../EventBus';
+import { EventBus } from '@casual-simulation/aux-components';
 import { AuxBotVisualizerFinder } from '../AuxBotVisualizerFinder';
 import { DebugObjectManager } from './debugobjectmanager/DebugObjectManager';
 import { AuxBot3D } from './AuxBot3D';
