+++ src/aux-server/aux-web/shared/scene/Text3D.ts
@@ -22,7 +22,7 @@ import {
 } from '@casual-simulation/aux-common';
 import { DebugObjectManager } from './debugobjectmanager/DebugObjectManager';
 import { Text as TextMesh } from 'troika-three-text';
-import Roboto from '../public/fonts/Roboto/roboto-v18-latin-regular.woff';
+import Roboto from '@casual-simulation/aux-components/fonts/Roboto/roboto-v18-latin-regular.woff';
 
 export interface Text3DFont {
     /**
